Implement maintaining scroll position when new squawka comes in
- implement a function that will memorize the position of the scroll position before the component re-renders
- implement a useLayoutEffect function that will set the scrollPostion if we saved any

FYI: the approach to this flatlist might be good for chats application. But if its for a Feed like a social media, you will want to appraoch this another way where you dont always re-render the flatlist. You allow the use read through the feeds you showed then without any obstruction. But you show then the laert that ther is a new feed. The when the click the alert thats when you re-render the whole component

Addtional likes on Flatlist configurations
- https://stackoverflow.com/questions/43397803/how-to-re-render-flatlist
- https://stackoverflow.com/questions/58907939/flatlist-scroll-to-top-while-flatlist-in-child-component
- https://github.com/facebook/react-native/issues/24206

https://stackoverflow.com/questions/29503252/get-current-scroll-position-of-scrollview-in-react-native (This links talks on things to have in maind about scroll position for iOS)

Additional resoures
https://stackoverflow.com/questions/38386704/react-native-determine-number-of-lines-of-text-component/39986135#39986135 (THis is more on shor more and show less)
- https://www.youtube.com/watch?v=0n_s1PLb_uY (lesson on animation on flatlist)